+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to `Livewire Combobox` will be documented in this file:
 ## 1.1 - 2021-05-04
 
 - New loading indicator.
+- New method `withoutResponse()`.
 - Fixed bug when main selector was reset.
 
 ## 1.0 - 2021-04-30